Stage_Test/terraform_test/requirements.txt
Flask==2.2.3
werkzeug==2.2.3
pandas==1.5.3
numpy==1.23.0
google-cloud-storage==2.10.0
setuptools>=58.0.0
wheel
google-cloud-firestore==2.10.0
